Move SerialPort and logger helpers to a new module
